// src/generated/model/searchRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FilterCriteria } from './filterCriteria.ts';
import type { FilterGroup } from './filterGroup.ts';
import type { Pagination } from './pagination.ts';
import type { SortCriteria } from './sortCriteria.ts';

export interface SearchRequest {
  filters?: FilterCriteria[];
  filterGroups?: FilterGroup[];
  sorts?: SortCriteria[];
  pagination?: Pagination;
}